feat: `ColonyContext` now provides a method to set connected colony
perf: `NetworkClient` is set up ahead of time to speed up load

refactor: split App component into two pages
